Improve solver tie-breaking and strengthen regression tests.
Refine compare_final with epsilon-based 2-ply tie-breaks and
turns-left-aware partition preference, align 2-ply follow-up scoring
with compare_one_ply, and deduplicate partition_sufficient. Add targeted
unit and integration tests plus a fast CI smoke suite. Fix README and
TUI key-binding docs, rename the suggestion metric label to Score, and
apply minor clippy cleanups without changing solve-rate benchmarks.

+++ src/tui/widgets.rs
@@ -62,7 +62,7 @@ impl Widget for TileRow<'_> {
         let start_x = area.x + (area.width.saturating_sub(total_width)) / 2;
 
         for i in 0..5 {
-            let x = start_x + (i as u16) * (cell_width + gap as u16);
+            let x = start_x + (i as u16) * (cell_width + gap);
             if x + cell_width > area.right() {
                 break;
             }
